Removed observer.OnCompleted()
OnCompleted() was called in Dispose(), but Dispose() is called only when
an observer is disconnecting. I believe OnCompleted should only be
called when the IServiceBus is shutting down, indicating that the
observable sequence will never see another value, but it doesn't look
like IServiceBus exposes an event for the bus shutdown.


Former-commit-id: b84c8186c6b1d6a6f0020908dc5b6726d59665d0