PSKESerializer SetContentType fix
Fixed an issue with regards to the
PreSharedKeyEncryptedMessageSerializer and MassTransit 2.7.4 which now
supports multiple serializers at Endpoint.Receive. The content type of
the serialized message was set to that of the wrapped serializer instead
of the encrypted serializer which ultimately caused a serialization
exception at the receiving endpoint. By setting the content type after
the wrapped serializer we now guarantee that the correct serializer is
selected.