Here lies MSMQ
Today we say good bye to some very old pieces of the MassTransit
code base. Pieces that were developed with love and passion.
But we have grown since then we've 'experimented' and found our
current love. It is RabbitMQ. MSMQ, its not you, its us. So smile
because it happened, don't cry because its over.