ArgumentException in InMemoryTransport.DispatchMessage
Not a very big problem since it's only in the InMemoryTransport, but something is weird.

I am not sure here what the correct behavior should be. The code check if `_receivePipe` is null, but
`_receivePipe` is initialized only by calling `IReceiveTransport.Start` and that is never called when retrieving the `ISendTransport` for the dead letter queue.

> This can be observed by running the `Publishing_a_message_on_the_bus_using_the_InMemoryTransport.Throw_an_an_uncatchable_ArgumentException`.

Additionally it's not clear to me why the message is moved to the dead letter queue even if an Handler is present.

> This can be observed by running the `Publishing_a_message_on_the_bus_using_the_InMemoryTransport.Should_not_move_the_message_to_the_dead_letter_queue`.

Note that the tests don't have any assertions, because i did not find a way to get hold of anything useful to demostrate the problem and i didn't want to make too many changes to the code without really understanding what's going on.

The only way to really observe the problem is by debugging the tests (watching the visual studio output windows will show a `Exception thrown: 'System.ArgumentException' in MassTransit.dll` after the message is published in both tests.
A proposed solution is available in *InMemoryTransport.cs* (commented out and marked with **L.P.**).