Update BinaryMessageSerializer.cs
We found an issue in one of our projects where we have a message that returns an Exception object as member, when the Exception has a inner member like in some Entity Framework exception that is not serializable it crash Masstransit because _formatter.Serialize keeps raising an exception and the message is never ACK to RabbitMQ then consumer keeps receiving the same message in an endless loop, creating dozen of requests per second and no message in _error queue.
Check IsSerializable doesn't produce the desired results if an inner member doesn't have IsSerializable.