Fix default ClientProvidedName
Previously, the default value of `ClientProvidedName`, which equals to
the process name, was never used since it was always being overwritten
with `null`. Now, the argument value is checked before the default value
is replaced.

In addition, remove the `connection_name` client property which will
automatically be set by the RabbitMQ.Client connection, using the value
passed to the `ConnectionFactory.CreateConnection` method.