Fix a typo in IBusControl.StartAsync XML doc
Also removed a comma. Seems to me that "Once the bus has been started it cannot be started again..." is a single statement and shouldn't be broken up by a comma. However the main purpose of this change is to fix the typo and I'm not hung up on the comma either way.

I didn't want to add the newline on the end of the file but the Github web editor seems to be inserting it automatically, if this is a deal breaker I'll fork and edit the file properly.