AmazonSQS transport: trying to get the queue and if does not exists creating it.
With those changes we do want to prevent provisioning new queues in the application, actual environment is already created with CloudFormation before the application is deployed.

With the previous implementation, when the applicatiion was not having CreateQueue permissions, it was never checking if the queue exists, as it was getting "AmazonSQSException: Access to the resource https://sqs.eu-west-1.amazonaws.com/ is denied." which was not handled in this code.