Limit JobTypeSaga concurrency by 1 thread.
This minimize amount of failures during version check of optimistic concurrency mechanism.
Good example is when we receive several Job Commands MT is creating JobSagas for each of them (in parallel) and tries to update JobTypeSaga singleton object with information about Active Jobs. In this case only the first update succeeds and all subsequent updates are failing on version check step. It is not critical, but is quite annoying and brings lots of errors in the log.