Add extra sqs permissions to sample iam policy
When creating an SQS queue and specifying tags to apply, the call to CreateQueue will fail with an AccessDenied error when using the original sample iam policy.
Internally, CreateQueue is calling TagQueue, but as that was missing from the sample policy, an AccessDenied error is thrown. There's no indication in any of the logging from AWS as to why this is though.

Checking the code shows that the SQS transport can also call Purge & Delete Queue as well, so have added those to avoid similar problems.