Update to ensure Marten Wire-Up works
Because we are outside of the generic host in the unit tests, the HostedService which Marten initialisation has moved to will never Start.

Also - This Internal HostedService (MartenActivator) has a dependency on an ILogger, so I've had to register one of those in the tests too by bringing in the MiIcrosoft.Extensions.Logging package.
I couldn't mock this as with ILogger of type T: MartenActivator, MartenActivator is an internal class.
I could have loaded the assembly, and obtained an instance of the class via reflection, but that felt a little "ew!" and bring in Microsoft.Extensions.Logging didn't feel too bad.

Other option would have been to roll a WebApplicationFactory and wire up the full pipeline, but that felt like it was overkill