4795 Throw Configuration Exception when using ConfigureEndpoints with Consumer, Saga, or Activity names (#4798)
Previously, naming a consumer class "Consumer" would result in an exception when building the bus. MassTransit removes the suffix "Consumer" when using the default configuration, leaving the consumer name an empty string. When this happens, a more descriptive exception is now thrown.

Clean up to simplify the exception and unit tests -pbg